Script can determine winner and play a round
The playRound() function asks for user input and prints out the result of the game to the console.

To determine the winner, the deterWinner function uses a switch function based on the users input. It returns a string indicating if the player won or not. If the user input was not correct, it prompts the user that they did not put in a valid input.