Updated JS file with keep score and end game func
The keep score function keeps track of the player and cpu score. Each round, it updates the game by removing the first child of the result container.

The end game function happens once the player or CPU has a total score of 5. The user will be prompt if they won or lost the game.